pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dense-rnn"
version = "0.1.0"
description = "Research code for DenseRNN experiments (copying, word problems, arithmetic, and language modeling)."
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
  "absl-py",
  "accelerate>=0.30",
  "dacite>=1.8",
  "datasets>=2.19",
  "einops>=0.7",
  "fire>=0.6",
  # Optional: flash-linear-attention requires torch>=2.5; moved to an extra to keep torch==2.4.1 workable
  "huggingface-hub>=0.23",
  "mamba-ssm==2.2.4",
  "mambapy>=0.1.0",
  "matplotlib>=3.8",
  "numpy>=1.24",
  "omegaconf>=2.3",
  "ordered-set>=4.1",
  "pandas>=2.0",
  "polars>=0.20",
  "pyyaml>=6.0",
  "scienceplots>=2.1",
  "schedulefree>=1.3",
  "scipy>=1.11",
  "seaborn>=0.13",
  "sympy>=1.12",
  "tokenizers>=0.15",
  "torch==2.4.1",
  "torchinfo>=1.8",
  "torchmetrics>=1.0.3,<1.1",
  "transformers>=4.42",
  "triton==3.0.0",
  "tqdm>=4.66",
  "wandb>=0.16",
  "abstract_algebra @ git+https://github.com/jopetty/abstract_algebra",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = [
  "plainLM",
  "plainLM.*",
  "copy_exp",
  "copy_exp.*",
  "word_problem_exp",
  "word_problem_exp.*",
  "arithmetic_exp",
  "arithmetic_exp.*",
  "models",
  "models.*",
  "data",
  "data.*",
]

[tool.setuptools.package-data]
"*" = ["*.yaml"]

[project.optional-dependencies]
conv = [
  # Optional: install `causal-conv1d` manually if you want the CUDA kernels; code falls back to PyTorch convs.
]
fla = []

[tool.uv]
extra-build-dependencies = { mamba-ssm = ["packaging", "wheel", "numpy", "torch==2.4.1", "nvidia-cuda-nvcc-cu12>=12.4.0,<12.5"] }

[dependency-groups]
dev = []
